fix more merge conflicts
